requirements.txt
fastapi
uvicorn
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
tinydb